badm_club_api/requirements.txt
asgiref>=3.8.1,<4
django-allauth==0.51.0
Pillow==9.4.0
sqlparse==0.4.3
djangorestframework==3.14.0
psycopg2==2.9.9
Django==5.1
openpyxl==3.1.5
gunicorn==23.0.0
whitenoise==6.11.0